ai_research/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
requests
beautifulsoup4
pandas
pandas-ta
yfinance
transformers
tenacity
apscheduler
python-multipart
python-slugify
